Refactor out MainFragment into its own class. Pass singleton dependencies to the new MainFragment class using dependency injection
